Journal Tab + Refactor
The journal tab is now complete and can fetch the relevant records via an API call.
CSS has been refactored and the faculty components have been organized as well.
There is now a loading and a records not found screen as well.
A common modal component has been created.
Additional information about a record can be viewed using a modal.