Journal Add + Conference Add + Refactor
Journal and Conference tabs now can add newer records.
They can inform the user if record is new, updated or if there is an error in the data.
The user delete, add and faculty add have been given to the director as well. This involved a refactor.
Minor migrations have been performed to allow blank value or 0 if integer.