Book Editorial Add + Book Chapter Add + View
Book editorial and Book chapters tabs now can add newer records.
They can inform the user if record is new, updated or if there is an error in the data.
The button has been moved to the filter tab which allows the user to use the button even when no records are present.
Minor migrations have been performed to allow blank value or 0 if integer.